+++ packages/launchpad/types/encoding.d.ts
@@ -22,12 +22,3 @@ export declare function makeStdSignDoc(
   sequence: number | string,
 ): StdSignDoc;
 export declare function serializeSignDoc(signDoc: StdSignDoc): Uint8Array;
-/** A convenience helper to create the StdSignDoc and serialize it */
-export declare function makeSignBytes(
-  msgs: readonly Msg[],
-  fee: StdFee,
-  chainId: string,
-  memo: string,
-  accountNumber: number | string,
-  sequence: number | string,
-): Uint8Array;